requirements.txt
praw==7.7.1
anthropic>=0.40.0
supabase==2.3.0
python-dotenv==1.0.0
